Authority refactor
