+++ lib/msal-core/src/ServerRequestParameters.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { Authority } from "./Authority";
+import { Authority } from "./authority/Authority";
 import { CryptoUtils } from "./utils/CryptoUtils";
 import { AuthenticationParameters, validateClaimsRequest } from "./AuthenticationParameters";
 import { StringDict } from "./MsalTypes";
