+++ lib/msal-core/src/authority/AuthorityFactory.ts
@@ -9,9 +9,9 @@
 import { AadAuthority } from "./AadAuthority";
 import { B2cAuthority } from "./B2cAuthority";
 import { Authority, AuthorityType } from "./Authority";
-import { ClientConfigurationErrorMessage } from "./error/ClientConfigurationError";
-import { UrlUtils } from "./utils/UrlUtils";
-import { StringUtils } from "./utils/StringUtils";
+import { ClientConfigurationErrorMessage } from "../error/ClientConfigurationError";
+import { UrlUtils } from "../utils/UrlUtils";
+import { StringUtils } from "../utils/StringUtils";
 
 export class AuthorityFactory {
     /**
@@ -24,8 +24,6 @@ export class AuthorityFactory {
         switch (pathSegments[0]) {
             case "tfp":
                 return AuthorityType.B2C;
-            case "adfs":
-                return AuthorityType.Adfs;
             default:
                 return AuthorityType.Aad;
         }
