+++ lib/msal-core/src/authority/ITenantDiscoveryResponse.ts
